TEK Sys coding challenge: - formatted all outputs

+++ app/src/main/java/test/app/teksysweather/model/Sys.java
@@ -20,8 +20,8 @@ public class Sys {
     public String country;
 
     @SerializedName("sunrise") @Expose
-    public int sunrise;
+    public long sunrise;
 
     @SerializedName("sunset") @Expose
-    public int sunset;
+    public long sunset;
 }
