+++ app/src/main/java/test/app/teksysweather/model/WeatherModel.java
@@ -39,7 +39,7 @@ public class WeatherModel {
     public Precipitation snow;
 
     @SerializedName("dt") @Expose
-    public int dt;
+    public long dt;
 
     @SerializedName("sys") @Expose
     public Sys sys;
